requirements.txt
cassandra-driver==3.17.0
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
six==1.12.0
Werkzeug==0.14.1
